Update ABR_GUI.m
Fixed the fact that an odd number of datatips was detected when changing from one axes to the other (the last one was removed because SHIFT has to be hold on).
Now the SavePoints function consider only the datatips in the current axes when checking for an even number of datatips. The last datatip from the previous axes still gets removed though.