Use a separate figure to perform peak analysis
To resolve some bugs encountered when using createdatatip (only for for versions earlier than Matlab 2019b), it's been decided to analyze peaks in a separate figure.
So a "Peak naalysis" button is replacing the peak detection panel, and plots the current ABR in a new figure where peak detection features will be added in future releases.